Bumps the bosh_aws_cpi gem 2.0.2 => 2.1
We also locked down the bosh_aws_cpi gem in all sub gemspecs owned by
the bosh project.  We can do this because there should be no consumers
of these gems anywhere that would need these dependencies to be looser,
and this exact version of the gem is the only version that we 'approve'
by running through our ci system.

[#104456530](https://www.pivotaltracker.com/story/show/104456530)

Signed-off-by: Beyhan Veli <beyhan.veli@sap.com>